+++ pkg/sysinfo/numcpu_linux.go
@@ -10,16 +10,6 @@ import (
 	"golang.org/x/sys/unix"
 )
 
-// Returns bit count of 1
-func popcnt(x uint64) (n byte) {
-	x -= (x >> 1) & 0x5555555555555555
-	x = (x>>2)&0x3333333333333333 + x&0x3333333333333333
-	x += x >> 4
-	x &= 0x0f0f0f0f0f0f0f0f
-	x *= 0x0101010101010101
-	return byte(x >> 56)
-}
-
 // numCPU queries the system for the count of threads available
 // for use to this process.
 //
